feat: rebuild project structure
- add db, models and api separation for the project
- reorganize and clean user routes and code
- create base files for upcoming models
- add enviroment variables parsing
